# bootstrapper/templates/bootstrapper.swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "A service to build bootstrap archives in various cloud environments. This\ntool will build S3 buckets in AWS, File Shares in Azure, Buckets in GCP, \ndatastore entries in VMWare and build ISO images for use with KVM. In short\nthis tool should help get VM-Series NGFWs up and running as quickly as \npossible. ",
    "version" : "0.1",
    "title" : "PAN-OS Bootstrap Utility"
  },
  "paths" : {
    "/bootstrap_azure" : {
      "post" : {
        "description" : "Bootstrap VM-Series on Microsoft Azure",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/bootstrap_azure_payload"
          },
          "x-examples" : {
            "application/json" : "{\n        \"azure_account_key\": \"zzzzzz\",\n        \"azure_account_name\": \"head1srvubuntudiag744\",\n        \"ADMINISTRATOR_USERNAME\": \"administrator\",\n        \"ADMINISTRATOR_PASSWORD\": \"asdfasdfasdf\",\n        \"bootstrap_template\": \"Default Bootstrap.xml\",\n        \"hostname\": \"panos-01\",\n        \"init_cfg_template\": \"Default Init-Cfg\",\n        \"dhcp_or_static\": \"dhcp_client\",\n        \"default_gateway\": \"\",\n        \"dg_name\": \"\",\n        \"dhcp_accept_server_domain\": \"\",\n        \"dhcp_accept_server_hostname\": \"\",\n        \"dhcp_send_client_id\": \"\",\n        \"dhcp_send_hostname\": \"\",\n        \"dns_primary\": \"\",\n        \"dns_secondary\": \"\",\n        \"ip_address\": \"\",\n        \"ipv6_address\": \"\",\n        \"ipv6_default_gateway\": \"\",\n        \"netmask\": \"\",\n        \"op_command_modes\": \"\",\n        \"panorama_server\": \"\",\n        \"panorama_server_2\": \"\",\n        \"tpl_name\": \"\",\n        \"vm_auth_key\": \"\"\n}"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Definition generated from Swagger Inspector"
          },
          "500" : {
            "description" : "Definition generated from Swagger Inspector"
          }
        }
      }
    },
    "/bootstrap_gcp" : {
      "post" : {
        "description" : "Bootstrap VM-Series on Google Compute Platform",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/bootstrap_gcp_payload"
          },
          "x-examples" : {
            "application/json" : "{\n        \"gcp_project_id\": \"possible-point-211418\",\n        \"gcp_access_token\": \"dfasdfasdf\",\n        \"ADMINISTRATOR_PASSWORD\": \"administrator\",\n        \"ADMINISTRATOR_USERNAME\": \"asdfasdfasdf\",\n        \"bootstrap_template\": \"Default Bootstrap.xml\",\n        \"hostname\": \"panos-01\",\n        \"init_cfg_template\": \"Default Init-Cfg\",\n        \"dhcp_or_static\": \"dhcp_client\",\n        \"default_gateway\": \"\",\n        \"dg_name\": \"\",\n        \"dhcp_accept_server_domain\": \"\",\n        \"dhcp_accept_server_hostname\": \"\",\n        \"dhcp_send_client_id\": \"\",\n        \"dhcp_send_hostname\": \"\",\n        \"dns_primary\": \"\",\n        \"dns_secondary\": \"\",\n        \"ip_address\": \"\",\n        \"ipv6_address\": \"\",\n        \"ipv6_default_gateway\": \"\",\n        \"netmask\": \"\",\n        \"op_command_modes\": \"\",\n        \"panorama_server\": \"\",\n        \"panorama_server_2\": \"\",\n        \"tpl_name\": \"\",\n        \"vm_auth_key\": \"\"\n}"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Definition generated from Swagger Inspector"
          }
        }
      }
    },
    "/bootstrap_aws" : {
      "post" : {
        "description" : "Bootstrap VM-Series on Amazon Web Services",
        "consumes" : [ "application/json" ],
        "produces" : [ "text/html" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/bootstrap_aws_payload"
          },
          "x-examples" : {
            "application/json" : "{\n        \"aws_secret\": \"abcdef\",\n        \"aws_key\": \"123456\",\n        \"aws_location\": \"us-east-2\",\n        \"ADMINISTRATOR_PASSWORD\": \"admin\",\n        \"ADMINISTRATOR_USERNAME\": \"youshouldchangeme\",\n        \"bootstrap_template\": \"Default Bootstrap.xml\",\n        \"hostname\": \"test-01\",\n        \"init_cfg_template\": \"Default Init-Cfg\",\n        \"dhcp_or_static\": \"dhcp_client\",\n        \"default_gateway\": \"\",\n        \"dg_name\": \"\",\n        \"dhcp_accept_server_domain\": \"}\",\n        \"dhcp_accept_server_hostname\": \"\",\n        \"dhcp_send_client_id\": \"\",\n        \"dhcp_send_hostname\": \"\",\n        \"dns_primary\": \"\",\n        \"dns_secondary\": \"\",\n        \"ip_address\": \"\",\n        \"ipv6_address\": \"\",\n        \"ipv6_default_gateway\": \"\",\n        \"netmask\": \"\",\n        \"op_command_modes\": \"\",\n        \"panorama_server\": \"\",\n        \"panorama_server_2\": \"\",\n        \"tpl_name\": \"\",\n        \"vm_auth_key\": \"\"\n}"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Definition generated from Swagger Inspector"
          }
        }
      }
    },
    "/bootstrap_kvm" : {
      "post" : {
        "description" : "Bootstrap VM-Series on Linux KVM. Returns an ISO image",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/iso-image" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/default_payload"
          },
          "x-examples" : {
            "application/json" : "{\n        \"ADMINISTRATOR_USERNAME\": \"administrator\",\n        \"ADMINISTRATOR_PASSWORD\": \"asdfasdfasdf\",\n        \"bootstrap_template\": \"Default Bootstrap.xml\",\n        \"hostname\": \"panos-01\",\n        \"init_cfg_template\": \"Default Init-Cfg\",\n        \"dhcp_or_static\": \"dhcp_client\",\n        \"default_gateway\": \"\",\n        \"dg_name\": \"\",\n        \"dhcp_accept_server_domain\": \"\",\n        \"dhcp_accept_server_hostname\": \"\",\n        \"dhcp_send_client_id\": \"\",\n        \"dhcp_send_hostname\": \"\",\n        \"dns_primary\": \"\",\n        \"dns_secondary\": \"\",\n        \"ip_address\": \"\",\n        \"ipv6_address\": \"\",\n        \"ipv6_default_gateway\": \"\",\n        \"netmask\": \"\",\n        \"op_command_modes\": \"\",\n        \"panorama_server\": \"\",\n        \"panorama_server_2\": \"\",\n        \"tpl_name\": \"\",\n        \"vm_auth_key\": \"\"\n}"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Definition generated from Swagger Inspector"
          },
          "500" : {
            "description" : "Definition generated from Swagger Inspector"
          }
        }
      }
    },
    "/get_bootstrap_variables" : {
      "post" : {
        "description" : "Return list of template variables for the given set of \nbootstrap templates\n",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/get_bootstrap_variables_payload"
          },
          "x-examples" : {
            "application/json" : "{\n    \"format\": \"aframe\",\n    \"bootstrap_template\": \"Default Bootstrap.xml\",\n    \"init_cfg_template\": \"Default Init-Cfg\",\n    \"authcodes_template\": \"None\"\n}"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Definition generated from Swagger Inspector"
          }
        }
      }
    },
    "/delete_template" : {
      "post" : {
        "description" : "Delete a template",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/delete_template_payload"
          },
          "x-examples" : {
            "application/json" : "{\n    \"template_name\": \"TEST_IMPORT\"\n}"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Definition generated from Swagger Inspector"
          }
        }
      }
    },
    "/import_template" : {
      "post" : {
        "description" : "Import a new template",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/import_template_payload"
          },
          "x-examples" : {
            "application/json" : "{\n    \"name\": \"TEST_IMPORT\",\n    \"description\": \"test import template\",\n    \"type\": \"bootstrap\",\n    \"template\": \"template string here with {{ var_name }}\"\n}"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Definition generated from Swagger Inspector"
          }
        }
      }
    },
    "/get_template" : {
      "post" : {
        "description" : "Return the text of a template",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/get_template_payload"
          },
          "x-examples" : {
            "application/json" : "{\n    \"template_name\": \"Default Bootstrap.xml\"\n}"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Definition generated from Swagger Inspector"
          }
        }
      }
    },
    "/render_template" : {
      "post" : {
        "description" : "\"Return the compiled template with all variables interpolated\"",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/text" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/render_template_payload"
          },
          "x-examples" : {
            "application/json" : "{\n    \"template_name\": \"TEST_IMPORT\",\n    \"var_name\": \"Some variable\"\n}"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Definition generated from Swagger Inspector",
            "x-examples" : {
              "application/text" : "template string here with Some variable"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "default_payload" : {
      "type" : "object",
      "properties" : {
        "ADMINISTRATOR_PASSWORD" : {
          "type" : "string"
        },
        "ADMINISTRATOR_USERNAME" : {
          "type" : "string"
        },
        "bootstrap_template" : {
          "type" : "string",
          "description" : "Name of the bootstrap template to use"
        },
        "hostname" : {
          "type" : "string"
        },
        "init_cfg_template" : {
          "type" : "string",
          "description" : "Name of the init-cfg template to use"
        },
        "dhcp_or_static" : {
          "type" : "string",
          "enum" : [ "dhcp_client", "static" ]
        },
        "default_gateway" : {
          "type" : "string"
        },
        "dg_name" : {
          "type" : "string"
        },
        "dhcp_accept_server_domain" : {
          "type" : "string"
        },
        "dhcp_accept_server_hostname" : {
          "type" : "string"
        },
        "dhcp_send_client_id" : {
          "type" : "string"
        },
        "dhcp_send_hostname" : {
          "type" : "string"
        },
        "dns_primary" : {
          "type" : "string"
        },
        "dns_secondary" : {
          "type" : "string"
        },
        "ip_address" : {
          "type" : "string"
        },
        "ipv6_address" : {
          "type" : "string"
        },
        "ipv6_default_gateway" : {
          "type" : "string"
        },
        "netmask" : {
          "type" : "string"
        },
        "op_command_modes" : {
          "type" : "string"
        },
        "panorama_server" : {
          "type" : "string"
        },
        "panorama_server_2" : {
          "type" : "string"
        },
        "tpl_name" : {
          "type" : "string"
        },
        "vm_auth_key" : {
          "type" : "string"
        }
      }
    },
    "bootstrap_azure_payload" : {
      "allOf" : [ {
        "$ref" : "#/definitions/default_payload"
      }, { } ]
    },
    "bootstrap_gcp_payload" : {
      "allOf" : [ {
        "$ref" : "#/definitions/default_payload"
      }, { } ]
    },
    "bootstrap_aws_payload" : {
      "allOf" : [ {
        "$ref" : "#/definitions/default_payload"
      }, { } ]
    },
...
    "get_bootstrap_variables_payload" : {
      "type" : "object",
      "properties" : {
        "format" : {
          "type" : "string"
        },
        "bootstrap_template" : {
          "type" : "string"
        },
        "init_cfg_template" : {
          "type" : "string"
        },
        "authcodes_template" : {
          "type" : "string"
        }
      }
    }
  }
}